expose at notification
