+++ at_client/test/samples/monitor/monitor_single_subscription_no_regex.dart
@@ -37,6 +37,6 @@ void main() async {
 //  }
 }
 
-void _notificationCallback(AtNotification notification) {
-  print('alice notification received: ${notification.toString()}');
-}
+//void _notificationCallback(AtNotification notification) {
+//  print('alice notification received: ${notification.toString()}');
+//}
