+++ at_client/test/samples/monitor/monitor_single_subscription_regex.dart
@@ -50,6 +50,6 @@ void main() async {
 //  print('end of test');
 }
 
-void _notificationCallback(AtNotification notification) {
-  print('alice notification received: ${notification.toString()}');
-}
+//void _notificationCallback(AtNotification notification) {
+//  print('alice notification received: ${notification.toString()}');
+//}
